requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
matplotlib